Turn on html5 mode.
A bug was found wherein a user was unable to render a promdash dashboard
within a frame widget if that dashboard relied on query string
parameters.

Without html5 mode engaged, the query string was only interpreted by the
angular code if it was preceded by a hash (#), e.g.:

http://promdash.com/my-dashboard/#/?some=components&and=others

However, the URL escaping code executed on frame URLs escapes all code
after the final hash (#), would would escape the '?' and break the
dashboard rendered in the frame.

With html5 mode engaged, the above URL can be written as one is
accustomed now:

http://promdash.com/my-dashboard?some=components&and=others